requirements.txt
wandb==0.15.11
tensorflow==2.8.0
streamlit==1.44.1
protobuf==3.20.3
pillow==10.0.1
pandas==2.1.1
numpy==1.26.0
Flask==3.0.2
click==8.1.3
